requirement.txt
annotated-types==0.6.0
antlr4-python3-runtime==4.12.0
anyio==4.3.0
argparse-dataclass==2.0.0
boto3==1.34.84
cachetools==5.3.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cmake==3.27.5
exceptiongroup==1.2.2
filelock==3.16.1
fsspec==2024.10.0
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.26.2
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
lit==16.0.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.127
overrides==7.7.0
packaging==24.2
pillow==11.0.0
plac==1.4.3
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.1.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pytest==8.3.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.10.1
safetensors==0.4.5
six==1.16.0
sniffio==1.3.1
spiral @ git+https://github.com/casics/spiral.git@0e75418cda7635f151fedff035b7b75799ccbe43
sympy==1.13.1
tdqm==0.0.1
tenacity==8.2.3
termcolor==2.5.0
tokenizers==0.20.3
tomli==2.1.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.5
transformers==4.46.3
triton==3.1.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
